% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causals.R
\name{build_design_matrix}
\alias{build_design_matrix}
\title{Builds the design matrix}
\usage{
build_design_matrix(data, beta)
}
\arguments{
\item{data}{a `data.frame` used to fit the FM-HighLD model}

\item{beta}{the estimated linear model coefficient. In the multi-trait case,
this is the estimated coefficient for the fixed effects}
}
\description{
Builds the design matrix
}
\examples{
set.seed(1234)
nsnps <- 20
data <- data.frame(var1 = rnorm(nsnps), var2 = rnorm(nsnps))
beta <- c("(Intercept)" = 2, "var1" = 1, "var2" = 2)
build_design_matrix(data, beta)
}
