% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{build_fm_tibble}
\alias{build_fm_tibble}
\title{Builds a `tibble::tibble` to be used by `FMHighLD`}
\usage{
build_fm_tibble(response, annot_matrix, ld_clusters, singletrait)
}
\arguments{
\item{response}{a vector for the `singletrait` case or a named list for
the multitrait case with an element for each variant. In the multitrait
case, each element of the list is a named vector where each element
is named after the trait for which the association was tested.}

\item{annot_matrix}{a matrix with same number of rows as variants and number
of columns as annotations}

\item{ld_clusters}{a character vector with the ld cluster to which each
variant belongs}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}
}
\value{
a `tibble::tibble` with at least columns snp, response, ld_cluster
 and the column names of `annot_matrix`. If `singletrait == FALSE`, it will
 also have a `trait column`.
}
\description{
Builds a `tibble::tibble` to be used by `FMHighLD`
}
\examples{
nsnps <- 10
z <- rlang::set_names(rnorm(nsnps), stringr::str_c("snp", seq_len(nsnps)))
annot <- matrix(rnorm(nsnps), ncol = 1)
rownames(annot) <- names(z)
colnames(annot) <- "annot"
ld_cluster <- sample(stringr::str_c("ld", seq_len(5)), nsnps, replace = TRUE)
names(ld_cluster) <- names(z)
build_fm_tibble(z, annot, ld_cluster, TRUE)
z_list <- list()
z_list[["a"]] <- z[1:3]
z_list[["b"]] <- z[3:5]
z_list[["c"]] <- z[6:10]
build_fm_tibble(z_list, annot, ld_cluster, FALSE)
}
