% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{build_formula}
\alias{build_formula}
\title{Utility to build the formula of the model of the underlying FM-HighLD model}
\usage{
build_formula(
  response,
  fixed_effects,
  fixed_intercept,
  random_effect_group = NULL,
  random_effects,
  random_intercept
)
}
\arguments{
\item{response}{name of the response variable}

\item{fixed_effects}{names of the fixed effects of the model}

\item{fixed_intercept}{logical indicating if there is a fixed effect
intercept, by default this values is `TRUE`}

\item{random_effect_group}{a character variable or NULL indicating if there
are random effects for the multi_trait case}

\item{random_effects}{names of the random effects of the model}

\item{random_intercept}{logical indicating if there is a random effect
intercept, by default this values is `TRUE`}
}
\description{
Utility to build the formula of the model of the underlying FM-HighLD model
}
\examples{
build_formula("z", c("skew", "var1", "var2"), FALSE)
build_formula("z", c("skew", "var1", "var2"), TRUE)
build_formula("z", c("skew", "var1", "var2"), TRUE, "gene", c("skew"), TRUE)
build_formula("z", c("skew", "var1", "var2"), TRUE, "gene", c("skew"), FALSE)
}
