% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_loglike_tibble}
\alias{build_loglike_tibble}
\title{Builds a likelihood list}
\usage{
build_loglike_tibble(iter_list)
}
\arguments{
\item{iter_list}{A list of iteraction after running `fmhighld_fit_[em|lm]`
with 'save_iter = TRUE'}
}
\value{
a `tibble::tibble`
}
\description{
Builds a likelihood list
}
