% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_variant_names}
\alias{check_variant_names}
\title{Checks that the variant names exists and are the same across the response,
 the annotation matrix and the ld_clusters vector}
\usage{
check_variant_names(response, annot_matrix, ld_clusters)
}
\arguments{
\item{response}{a vector for the `singletrait` case or a named list for
the multitrait case with an element for each variant. In the multitrait
case, each element of the list is a named vector where each element
is named after the trait for which the association was tested.}

\item{annot_matrix}{a matrix with same number of rows as variants and number
of columns as annotations}

\item{ld_clusters}{a character vector with the ld cluster to which each
variant belongs}
}
\value{
a logical indicator determining if the response, annotation matrix
 and ld_clusters variables are named, and whether the names are the same
}
\description{
Checks that the variant names exists and are the same across the response,
 the annotation matrix and the ld_clusters vector
}
