% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{compute_ith_model}
\alias{compute_ith_model}
\title{Computes the i-th mixture model for fitting the FM-HighLD model}
\usage{
compute_ith_model(
  i,
  causal_candidates,
  formula,
  data,
  gamma,
  singletrait = TRUE
)
}
\arguments{
\item{i}{a numeric index, usually i = 2 only that correspond to the causal
variants, but the model is performed for a list in case more models are
being fitted for the locus}

\item{causal_candidates}{a vector of the causal candidates selected for each
LD cluster}

\item{formula}{formula used for fitting the underlying linear model used by
FM-HighLD}

\item{data}{`data.frame` used for the FM-HighLD model}

\item{gamma}{a `Matrix::Matrix` with the conditional probabilities of the
variants being in each model}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}
}
\description{
Computes the i-th mixture model for fitting the FM-HighLD model
}
