% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{compute_mixture_prob}
\alias{compute_mixture_prob}
\title{Compute the mixture probabilities from a probability matrix}
\usage{
compute_mixture_prob(prob_matrix)
}
\arguments{
\item{prob_matrix}{a matrix of dimensions nassoc x nmixtures}
}
\value{
a vector of length nmixtures
}
\description{
Compute the mixture probabilities from a probability matrix
}
\examples{
compute_mixture_prob(matrix(rep(1, 100), ncol = 2))
}
