% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_iteration_single}
\alias{em_iteration_single}
\title{Performs an Expectation-Maximization iterations for the single
trait model}
\usage{
em_iteration_single(
  formula,
  data,
  causal_current,
  prev_iter,
  sigma0,
  fm_param,
  verbose
)
}
\arguments{
\item{formula}{a formula object with the underlying linear model}

\item{data}{a `data.frame` object with the annotation matrix and z-values}

\item{causal_current}{A list of `tibble::tibble` with columns `ld_cluster`
and `which_snp` obtained in the current iteration of the algorithm}

\item{prev_iter}{a `FMIter` object obtained in the previous iteration of the
FMHighLD algorithm}

\item{sigma0}{the error variance estimate of the background model}

\item{fm_param}{configuration parameter used to pick the causal candidate
per group}

\item{verbose}{a logical indicator of adding process messages}
}
\value{
a list with the updated parameters in one EM iteration
}
\description{
Performs an Expectation-Maximization iterations for the single
trait model
}
