% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_causal_vector}
\alias{extract_causal_vector}
\title{extract causal candidate vectors}
\usage{
extract_causal_vector(causal_list)
}
\arguments{
\item{causal_list}{a list of `tibble::tibble` with columns ld_cluster and
which_snp}
}
\value{
a list of the same length as `causal_list` with two names character
 vectors with the causal candidate SNPs
}
\description{
extract causal candidate vectors
}
\examples{
causal_list <- list(tibble::tibble(
 ld_cluster = stringr::str_c("ld", 1:2),
 which_snp = stringr::str_c("snp", 1:2)))
extract_causal_vector(causal_list)
}
