% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{flexmix_fit}
\alias{flexmix_fit}
\title{Fits the EM algorithm part of the FMHighLD model using flexmix}
\usage{
flexmix_fit(formula, data, causal_candidates, ncausal_mixt, ...)
}
\arguments{
\item{formula}{Formula of the non-zero mean part of the FMHighLD model}

\item{data}{`data.frame` used for the FM-HighLD model}

\item{causal_candidates}{a list with vectors of causal candidates}

\item{ncausal_mixt}{number of components in the mixture model}

\item{...}{other parameters using to fit `flexmix::flexmix`}
}
\value{
a `flexmix::flexmix` object
}
\description{
Fits the EM algorithm part of the FMHighLD model using flexmix
}
