% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FMDataSet-class}
\alias{FMDataSet-class}
\alias{FMDataSet}
\alias{FMDataSet-class}
\title{FMDataSet object and constructors}
\usage{
FMDataSet(snp_data, annot_conf, response, LD_matrix)
}
\arguments{
\item{snp_data}{a data.frame with at least the columns SNP, seqnames and position. 
The remaining columns are going to be considered as annotations.}

\item{annot_conf}{a data.frame with the configuration of the annotation data as fixed or 
random effects.}

\item{response}{a data.frame with three columns SNP, pheno and the value of an association 
statistic between both. For example, for a continuous phenotype we would use the wald statistic.}

\item{LD_matrix}{a matrix with the correlation between all the SNPs in snp_data.}
}
\value{
A FMDataSet object.
}
\description{
\code{FMDataSet} is a a subclass of \code{GenomicRanges}, used to store the 
input values, intermediate calculations and results of a multi-response fine-mapping analysis 
of SNPs in High LD.
}
\examples{
snps =c("a","b","c","d","e","f")
nsnps = length(snps)
nassoc = 20
snp_data = data.frame(SNP = snps,
                  seqnames = rep("chrA",nsnps),
                  position = sample(1e6,nsnps),
                  annot1 = factor(sample(c(TRUE,FALSE),nsnps,replace = TRUE )),
                  annot2 = rnorm(nsnps))
annot_conf = data.frame(annot = c("annot1","annot2"),
                        fixed = c(TRUE,TRUE),random = c(TRUE,FALSE))
response = data.frame(SNP = sample(snps,nassoc,replace = TRUE),
                     pheno = sample(c("I","II","III"),nassoc,replace = TRUE),
                     z = rnorm(nassoc,sd = 30))
LD_matrix = matrix(runif(nsnps^2),nrow = nsnps)
LD_matrix = .5 * (LD_matrix + t(LD_matrix))

fm = FMDataSet(snp_data,annot_conf,response,LD_matrix)
fm
}
