% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fmhighld_fit}
\alias{fmhighld_fit}
\title{Fit the FMHighLD model}
\usage{
fmhighld_fit(
  response,
  annot_matrix,
  ld_clusters,
  singletrait = TRUE,
  ncausal_mixt = 1,
  formula = NULL,
  skip_causal = FALSE,
  fm_param = FMParam(),
  save_iter = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{response}{a vector for the `singletrait` case or a named list for
the multitrait case with an element for each variant. In the multitrait
case, each element of the list is a named vector where each element
is named after the trait for which the association was tested.}

\item{annot_matrix}{a matrix with same number of rows as variants and number
of columns as annotations}

\item{ld_clusters}{a character vector with the ld cluster to which each
variant belongs}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}

\item{ncausal_mixt}{the number of mixtures used in the causal models}

\item{formula}{a `formula` for the underlying linear models, by default is
`NULL`, in which case `fmhighld_fit` will create the formula.}

\item{skip_causal}{a logical indicator determining whether the function will
skip the causal selection steps, and only iterate through the EM algorithm}

\item{fm_param}{a `FMParam` object with the parameters used to run `FMHighLD`}

\item{save_iter}{a logical indicator determining whether the iteration data
is going to be returned}

\item{verbose}{a logical indicator determining whether messages are going to
be used}
}
\value{
results
}
\description{
Fit the FMHighLD model
}
