% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FMIter-class}
\alias{FMIter-class}
\alias{FMIter}
\title{`FMIter` object}
\usage{
FMIter(
  nassoc = NULL,
  singletrait = NULL,
  models = NULL,
  causal_candidates = NULL,
  gamma = NULL,
  mu = NULL,
  sigma = NULL
)
}
\arguments{
\item{nassoc}{a numeric variable with the number of association tested
in the experiment}

\item{singletrait}{a logical indicator determininig if we are considering a
single trait or multiple traits}

\item{models}{a list with the models in the mixture}

\item{causal_candidates}{a list of the same length as `models` made with
vectors of causal candidate snps per each valid trait - ld cluster
combination}

\item{gamma}{a matrix with the posterior probabilities of a snps belonging to
any element in the mixture}

\item{mu}{a matrix with the predicted values of every snp with each column
being an element in the mixture}

\item{sigma}{a matrix with the variances of every normal distribution in the
mixture.}
}
\description{
`FMIter` contains the elements of an iteration for the FMHighLD algorithm
}
