% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_FMParam.R
\docType{methods}
\name{strategy}
\alias{strategy}
\alias{params_all}
\alias{params_pickm}
\alias{error_bound}
\alias{max_iter}
\alias{min_tol}
\alias{annot_tol}
\alias{show,FMParam-method}
\alias{show}
\alias{strategy,FMParam-method}
\alias{params_all,FMParam-method}
\alias{params_pickm,FMParam-method}
\alias{error_bound,FMParam-method}
\alias{max_iter,FMParam-method}
\alias{min_tol,FMParam-method}
\alias{annot_tol,FMParam-method}
\title{strategy method}
\usage{
strategy(object)

params_all(object)

params_pickm(object)

error_bound(object)

max_iter(object)

min_tol(object)

annot_tol(object)

\S4method{show}{FMParam}(object)

\S4method{strategy}{FMParam}(object)

\S4method{params_all}{FMParam}(object)

\S4method{params_pickm}{FMParam}(object)

\S4method{error_bound}{FMParam}(object)

\S4method{max_iter}{FMParam}(object)

\S4method{min_tol}{FMParam}(object)

\S4method{annot_tol}{FMParam}(object)
}
\arguments{
\item{object}{a `FMParam` object}
}
\value{
The randomization strategy to be used by `FMHighLD`

The randomization probability to be used by `FMHighLD` and the
absolute residual rank to select another causal candidate

The randomization probability to be used by `FMHighLD` and the
absolute residual rank to select another causal candidate

the error bound

the max. number of iterations

the min. change tolerance used by the FMHighLD algorithm
}
\description{
`strategy` returns the randomization strategy

`params_all` returns a list with the randomization probability to be used
by `FMHighLD` and the  absolute residual rank to select another causal
candidate

`params_pickm` returns a list with the randomization probability to
be used by `FMHighLD`, the  absolute residual rank to select another
causal candidate, the number of LD groups to apply the randomization
strategy and the method used to pick those LD groups

`error_bound` returns a constant indicating the min. values used to avoid
zero denominators

`max_iter` returns the maximum number of iteration used by the FMHighLD
algorithm

`min_tol` returns the minimum tolerance used by the FMHighLD
algorithm

`annot_tol` returns the minimum tolerance to accept variants with annotation
equal to zero
}
\examples{
strategy(FMParam())
params_all(FMParam(strategy = "all", prob = .9, k = 2))
params_pickm(FMParam(strategy = "pick_m", prob = .9, k = 2))
error_bound(FMParam())
max_iter(FMParam())
min_tol(FMParam())
annot_tol(FMParam())
}
\references{
the min. tolerance used to accept an annotation equal to zero
}
