% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FMParam-class}
\alias{FMParam-class}
\alias{FMParam}
\title{Constructor for the `FMParam` class}
\usage{
FMParam(
  strategy = NULL,
  prob = NULL,
  k = NULL,
  m = NULL,
  m_sel = NULL,
  error_bound = 1e-04,
  max_iter = 100,
  min_tol = 1e-06
)
}
\arguments{
\item{strategy}{name for the randomization strategy to be used. It could be
one of `none`, `all` or `pick_m`. By default, if all parameters of the
function are `NULL` the strategy is going to be `none`.
For the `all` strategy, it will pick the the `k`-th variant (in terms of
their absolute(residual) with probability `prob`. Finally, for the `pick_m`
strategy, the `FMHighLD` algorithm will apply a similar strategy but for `m`
ld groups. The `m` ld groups will be selected randomly if `m_strat` is equal
to any, or like the top `m` ld groups with highest average LD score}

\item{prob}{probability used to randomize the causal candidate, by default
it is 0.95}

\item{k}{absolute residual index of the causal candidate to select with
probability `prob`.  The default this strategy will selected the variant with
the second rank}

\item{m}{number of ld groups for m_sel the randomization strategy is going to
be applied. By default, it will apply the strategy to only one ld group.}

\item{m_sel}{Selection strategy to pick the `m` ld groups to be randomized.
By default, it will pick the `m` groups randomly. Alternatively, it could
also select the `m` groups by their highest average ld score. Either one of
`random` or `large_ld`.}

\item{error_bound}{a double constant indicating the min value used to avoid
zero denominators}

\item{max_iter}{the max. number of iteration by the algorithm used to stop
the algorithm}

\item{min_tol}{the min. tolerance used to stop the algorithm}
}
\value{
A `FMParam` object with the `FMHighLD` algorithm configuration
}
\description{
`FMParam` contains the configuration for randomized selection of causal
candidate variants
}
