% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_FMRandParam.R
\docType{methods}
\name{strategy}
\alias{strategy}
\alias{params_all}
\alias{params_pickm}
\alias{strategy,FMRandParam-method}
\alias{params_all,FMRandParam-method}
\alias{params_pickm,FMRandParam-method}
\title{strategy method}
\usage{
strategy(object)

params_all(object)

params_pickm(object)

\S4method{strategy}{FMRandParam}(object)

\S4method{params_all}{FMRandParam}(object)

\S4method{params_pickm}{FMRandParam}(object)
}
\arguments{
\item{object}{A `FMRandParam` object.}
}
\value{
The randomization strategy to be used by `FMHighLD`

The randomization probability to be used by `FMHighLD` and the
absolute residual rank to select another causal candidate

The randomization probability to be used by `FMHighLD` and the
absolute residual rank to select another causal candidate
}
\description{
`strategy` returns the randomization strategy

`params_all` returns a list with the randomization probability to be used
by `FMHighLD` and the  absolute residual rank to select another causal
candidate

`params_pickm` returns a list with the randomization probability to
be used by `FMHighLD`, the  absolute residual rank to select another
causal candidate, the number of ld groups to apply the randomization
strategy and the method used to pick those ld groups
}
