% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{get_ld_clusters}
\alias{get_ld_clusters}
\title{Get the LD clusters out of an LD matrix}
\usage{
get_ld_clusters(ld_matrix, min_r2, snps = NULL, max_dist = Inf, tidy = FALSE)
}
\arguments{
\item{ld_matrix}{a square matrix with the squared correlation between every
pair of SNPs in a locus.}

\item{min_r2}{min. R^2 for two SNPs to be in the same LD cluster}

\item{snps}{a named position vector, `IRanges::IRanges` or
`GRanges::GenomicRanges` object to account for genomic distance between
SNPs when clustering them.}

\item{max_dist}{max. distance between SNPs. When the SNPs vector is `NULL`.
The function will assume that the distance is one between adjacent SNPs.}

\item{tidy}{a logical indicator to return a `tibble::tibble` with `snp` and
`cluster` columns}
}
\value{
a membership LD cluster vector
}
\description{
Get the LD clusters out of an LD matrix
}
\examples{
set.seed(123)
rho_hub = c(.9, .7)
tau_hub = c( (.9 - .7) / (10-2), (.7-.6)/(5-2))
eps_hub = min(1-(rho_hub) - 0.75*(tau_hub) ) - .01
corrmat <- sim_hub_cor(block_sizes = c(10, 5),
 max_corrs = rho_hub, min_corrs = c(.7,.6),
 power = 1, epsilon = eps_hub, eidim=2)
get_ld_clusters(corrmat^2, .8)
rownames(corrmat) <- colnames(corrmat) <- letters[seq_len(nrow(corrmat))]
get_ld_clusters(corrmat^2, .8)
position <- sort(runif(nrow(corrmat), min = 10, max = 15))
names(position) <- letters[seq_along(position)]
get_ld_clusters(corrmat^2, .8, snps = position, max_dist = .5)
position <- IRanges::IRanges(start = position, width = 1)
names(position) <- letters[seq_along(position)]
get_ld_clusters(corrmat^2, .8, snps = position, max_dist = .5, tidy = TRUE)
position <- GenomicRanges::GRanges(seqnames = "chrom", ranges = position)
get_ld_clusters(corrmat^2, .8, snps = position, max_dist = .3, tidy = TRUE)
}
