% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{get_response_name}
\alias{get_response_name}
\title{Gets the name of the response based on the formula}
\usage{
get_response_name(formula)
}
\arguments{
\item{formula}{a formula object with the underlying linear / linear mixed
model}
}
\value{
the name of the response
}
\description{
Gets the name of the response based on the formula
}
