% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_snp_names}
\alias{get_snp_names}
\title{Gets a sorted vector with the snp names}
\usage{
get_snp_names(response, singletrait)
}
\arguments{
\item{response}{Either a numeric vector when `singletrait == TRUE` or a
list with named vectors when `singletrait == FALSE`}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}
}
\value{
a sorted character vector with the names of all the snps in the
  response
}
\description{
Gets a sorted vector with the snp names
}
\examples{
nsnps <- 10
z <- rlang::set_names(rnorm(nsnps), stringr::str_c("snp", seq_len(nsnps)))
get_snp_names(z, TRUE)
z_list <- list()
z_list[["a"]] <- z[1:3]
z_list[["b"]] <- z[3:5]
z_list[["c"]] <- z[6:10]
get_snp_names(z_list,  FALSE)
}
