% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{get_unif_block_sizes}
\alias{get_unif_block_sizes}
\title{Gets the LD block structure given the number of snps and number of LD
clusters.}
\usage{
get_unif_block_sizes(nsnps, nldblocks)
}
\arguments{
\item{nsnps}{Number of SNPs to simulate}

\item{nldblocks}{Number of LD blocks to simulate}
}
\value{
A vector with the number of snps per ld block, if the number nsnps
is not a multiple of nldblocks if will complete with the remaining number of
snps
}
\description{
Gets the LD block structure given the number of snps and number of LD
clusters.
}
\examples{
get_unif_block_sizes(50, 10)
get_unif_block_sizes(50, 11)
}
