% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_intercept}
\alias{has_intercept}
\title{Checks whether the formula contains the intercept}
\usage{
has_intercept(formula)
}
\arguments{
\item{formula}{a lm / lmm `formula` for the underlying linear models}
}
\value{
a logical value indicating whether the fixed effects have an
 intercept term
}
\description{
Checks whether the formula contains the intercept
}
