% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causals.R
\name{include_intercept}
\alias{include_intercept}
\title{Checks whether the intercept is include in the underlying linear model}
\usage{
include_intercept(features)
}
\arguments{
\item{features}{names of the estimated coefficient}
}
\value{
a logical indicating whether include the intercept in the model
}
\description{
Checks whether the intercept is include in the underlying linear model
}
