% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_causal_candidates}
\alias{init_causal_candidates}
\title{Initializes the causal candidates before fitting the model}
\usage{
init_causal_candidates(data, singletrait, ncausal_mixt)
}
\arguments{
\item{data}{`data.frame` used for the FM-HighLD model}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}

\item{ncausal_mixt}{the number of mixtures used in the causal models}
}
\value{
a list of length `ncausal_mixt` with a vector of one causal candida-
 te per ld group (if `singletrait = TRUE`) or per combination of ld_group
 and trait if (`singletrait = FALSE`)
}
\description{
Initializes the causal candidates before fitting the model
}
\examples{
nsnps <- 10
z <- rlang::set_names(rnorm(nsnps), stringr::str_c("snp", seq_len(nsnps)))
annot <- matrix(rnorm(nsnps), ncol = 1)
rownames(annot) <- names(z)
colnames(annot) <- "annot"
ld_cluster <- sample(stringr::str_c("ld", seq_len(5)), nsnps, replace = TRUE)
names(ld_cluster) <- names(z)
data <- build_fm_tibble(z, annot, ld_cluster, TRUE)
init_causal_candidates(data, TRUE, 1)
init_causal_candidates(data, TRUE, 2)
z_list <- list()
z_list[["a"]] <- z[1:3]
z_list[["b"]] <- z[3:5]
z_list[["c"]] <- z[6:10]
data = build_fm_tibble(z_list, annot, ld_cluster, FALSE)
init_causal_candidates(data, FALSE, 3)
}
