% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_causal_candidates_coef}
\alias{init_causal_candidates_coef}
\title{Initializes the causal candidates before fitting the model by picking the
snps based on a model with sampled random coefficients from a normal
distribution centered around the `init_coef_mean`}
\usage{
init_causal_candidates_coef(
  data,
  singletrait,
  formula,
  init_coef_mean,
  ncausal_mixt,
  fm_param
)
}
\arguments{
\item{data}{`data.frame` used for the FM-HighLD model}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}

\item{formula}{the formula of the underlying linear model}

\item{init_coef_mean}{the mean vector to sample the init. coefficients from
a normal distribution}

\item{ncausal_mixt}{the number of mixtures used in the causal models}

\item{fm_param}{a `FMParam` object with the parameters used to run `FMHighLD`}
}
\value{
a list of length `ncausal_mixt` with a vector of one causal candida-
 te per ld group (if `singletrait = TRUE`) or per combination of ld_group
 and trait if (`singletrait = FALSE`)
}
\description{
Initializes the causal candidates before fitting the model by picking the
snps based on a model with sampled random coefficients from a normal
distribution centered around the `init_coef_mean`
}
\examples{
nsnps <- 10
z <- rlang::set_names(rnorm(nsnps), stringr::str_c("snp", seq_len(nsnps)))
annot <- matrix(rnorm(nsnps), ncol = 1)
rownames(annot) <- names(z)
colnames(annot) <- "annot"
ld_cluster <- sample(stringr::str_c("ld", seq_len(5)), nsnps, replace = TRUE)
names(ld_cluster) <- names(z)
data <- build_fm_tibble(z, annot, ld_cluster, TRUE)
init_causal_candidates_coef(data, TRUE, response ~ 0 + annot, .5, 1,
 FMParam())
init_causal_candidates_coef(data, TRUE, response ~ 0 + annot, .5, 2,
 FMParam())
}
