% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R, R/fit_model_em.R
\name{init_iteration_em}
\alias{init_iteration_em}
\title{Performs the initial iteration of the FM-HighLD model}
\usage{
init_iteration_em(formula, data, singletrait, ncausal_mixt, fm_param)

init_iteration_em(formula, data, singletrait, ncausal_mixt, fm_param)
}
\arguments{
\item{formula}{formula used for fitting the underlying linear model used by
FM-HighLD}

\item{data}{`data.frame` used for the FM-HighLD model}

\item{singletrait}{a logical indicator determining if the model is for
multi_trait or single-trait fine-mapping}

\item{ncausal_mixt}{the number of mixtures used in the causal models}

\item{fm_param}{a `FMParam` object with the parameters used to run `FMHighLD`}
}
\value{
a list with the models, and probability matrix for the EM-algorithm

a list with the models, and probability matrix for the EM-algorithm
}
\description{
Performs the initial iteration of the FM-HighLD model

Performs the initial iteration of the FM-HighLD model
}
