% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jsd}
\alias{jsd}
\title{Computes the Jensen-Shannon (JS) divergence between two discrete
probability distributions}
\usage{
jsd(p1, p2)
}
\arguments{
\item{p1}{the first probability vector}

\item{p2}{the second probability vector}
}
\value{
a scalar value with the JS-divergence between both
 distributions
}
\description{
Computes the Jensen-Shannon (JS) divergence between two discrete
probability distributions
}
\examples{

u1 <- runif(100)
u1 <- u1 / sum(u1)
v1 <- runif(100)
v1 <- v1 / sum(v1)
jsd(u1, v1)
}
