% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kl}
\alias{kl}
\title{Compute the Kullback-Leibler (KL) divergence between two discrete
probability distributions}
\usage{
kl(p1, p2)
}
\arguments{
\item{p1}{the first probability vector}

\item{p2}{the second probability vector}
}
\value{
a scalar value with the KL-divergence between both
 distributions
}
\description{
Compute the Kullback-Leibler (KL) divergence between two discrete
probability distributions
}
\examples{

u1 <- runif(100)
u1 <- u1 / sum(u1)
v1 <- runif(100)
v1 <- v1 / sum(v1)
kl(u1, v1)
}
