% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lm_std_err}
\alias{lm_std_err}
\title{Gets the std. error from a linear model}
\usage{
lm_std_err(lm_model)
}
\arguments{
\item{lm_model}{an `lm` object}
}
\value{
the std. error of `lm_model`
}
\description{
Gets the std. error from a linear model
}
