% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmm_comp_var}
\alias{lmm_comp_var}
\title{Computes the variance of a linear mixed model}
\usage{
lmm_comp_var(var_factors, residual_error, re_error)
}
\arguments{
\item{var_factors}{a vector with the variances for each component of
the linear mixed model random effects}

\item{residual_error}{a vector with the residual errors of the
linear mixed model used by FM-HighLD}

\item{re_error}{a vector with the random effect erros of the linear
mixed model used by FM-HighLD}
}
\value{
The variance matrix of the model
}
\description{
Computes the variance of a linear mixed model
}
