% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{noise_cor}
\alias{noise_cor}
\title{adds noise to a user specific matrix}
\usage{
noise_cor(corr_matrix, epsilon = 0.01, eidim = 2)
}
\arguments{
\item{corr_matrix}{a pre-defined correlation matrix}

\item{epsilon}{maximum entry wise random noise}

\item{eidim}{dimension of the noise space, the smaller the value gets more
noise}
}
\value{
a simulated correlation matrix
}
\description{
adds noise to a user specific matrix
}
\note{
Based on Hardin et al, "A method for generating realistic
correlation matrices", Annals of Applied Statistics (2012)
}
\examples{
noise_cor(diag(15), epsilon = .5, eidim=2)
}
