% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm2_wrap}
\alias{norm2_wrap}
\title{Wrap around the `norm` operator to work with `dplyr::rowwise`}
\usage{
norm2_wrap(...)
}
\arguments{
\item{...}{names of the variables to use when computing the norm2}
}
\value{
the norm2 of the columns picks
}
\description{
Wrap around the `norm` operator to work with `dplyr::rowwise`
}
\examples{
library(magrittr)
tibble::tribble( ~ "vec1", ~ "vec2", 1, 1, 0, 1, 1, 0, 0, 0) \%>\%
 dplyr::rowwise() \%>\%
 dplyr::mutate( norm2_wrap(vec1, vec2))
}
