% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{parse_re_variance}
\alias{parse_re_variance}
\title{Parses the random error variance matrix}
\usage{
parse_re_variance(re_error, ntraits)
}
\arguments{
\item{re_error}{estimates of the RE errors}

\item{ntraits}{integer with the number of traits}
}
\value{
a sparse `Matrix` object
}
\description{
Parses the random error variance matrix
}
