% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{philips}
\alias{philips}
\title{Computes the philips stopping criteria across all distances}
\usage{
philips(x)
}
\arguments{
\item{x}{a vector of losses between current and previous iterations}
}
\value{
the philips stopping criteria
}
\description{
Computes the philips stopping criteria across all distances
}
