% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{rho_hub_aux}
\alias{rho_hub_aux}
\title{Auxiliary function to fill in the rest of the structure of the Hub
correlation matrix}
\usage{
rho_hub_aux(r_max, r_min, power, p)
}
\arguments{
\item{r_max}{is the maximum user specified correlation}

\item{r_min}{is the minimum user specified correlation}

\item{power}{is the power at which the correlations descend}

\item{p}{is the size of the correlation block}
}
\description{
Auxiliary function to fill in the rest of the structure of the Hub
correlation matrix
}
