% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causals.R
\name{select_causals_multi}
\alias{select_causals_multi}
\title{Selects the causal variants in the locus for a multi-trait model}
\usage{
select_causals_multi(data, model, group, rand_param = NULL, cond_res = FALSE)
}
\arguments{
\item{data}{a `data.frame` used to fit the FM-HighLD model}

\item{model}{a fitted model with the previously selected causal candidates}

\item{group}{name of the grouping variable used to pick the causal
candidates}

\item{rand_param}{configuration parameter used to pick the caI usal candidate
per group}

\item{cond_res}{a logical indicating whether to use conditional residuals
of the linear mixed model}
}
\description{
Selects the causal variants in the locus for a multi-trait model
}
