% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causals.R
\name{select_kth_random}
\alias{select_kth_random}
\title{Selects the k-th smallest squared residual randomly with probability
 prob}
\usage{
select_kth_random(res, prob = 0.5, k = 2)
}
\arguments{
\item{res}{a vector of residuals}

\item{prob}{a probability to randomize the selection of the k-th
smallest residual}

\item{k}{an integer value}
}
\value{
the id of the selected k-th smallest residual
}
\description{
Selects the k-th smallest squared residual randomly with probability
 prob
}
