% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{sim_block_cor}
\alias{sim_block_cor}
\title{Simulates a block correlation matrix with constant correlation per block
where the size and base correlation for each block is user specified}
\usage{
sim_block_cor(
  block_sizes = c(10, 5, 8, 2, 15, 50),
  corrs = c(0.7, 0.7, 0.5, 0.9, 0.85, 0.4),
  delta = 0.39,
  epsilon = 0.99 - max(corrs),
  eidim = 2
)
}
\arguments{
\item{block_sizes}{A vector with the dimensions of each block}

\item{corrs}{A vector of same dimension than `block_sizes` with the base
correlation for each block}

\item{delta}{An additional parameter for the off diagonal correlations}

\item{epsilon}{maximum entry wise random noise}

\item{eidim}{dimension of the noise space, the smaller the value gets more
noise}
}
\value{
a simulated correlation matrix
}
\description{
Simulates a block correlation matrix with constant correlation per block
where the size and base correlation for each block is user specified
}
\note{
Based on Hardin et. al., "A method for generating realistic
correlation matrices", Annals of Applied Statistics (2012)
}
\examples{
rho_block <- c(.9, .2)
eps_block <- 0.99 - max(rho_block)
sim_block_cor(
 block_sizes = c(10, 5), corrs = rho_block, delta = 0.39,
 epsilon = eps_block, eidim = 2)
}
