% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{sim_hub_cor}
\alias{sim_hub_cor}
\title{Calculates a Toeplitz matrix with values descending from a user specified
range of correlations per block. The matrix has a block diagonal structure
and the size and base correlation for each block is user specified.}
\usage{
sim_hub_cor(
  block_sizes = c(10, 5, 8, 7, 15, 50),
  max_corrs = c(0.9, 0.7, 0.7, 0.5, 0.9, 0.3),
  min_corrs = c(0.7, 0.7, 0.2, 0.3, 0.85, 0.2),
  power = 1,
  epsilon = 0.08,
  eidim = 2
)
}
\arguments{
\item{block_sizes}{A vector with the dimensions of each block}

\item{max_corrs}{A vector of same dimension than `block_sizes` with the max
correlation for each block}

\item{min_corrs}{A vector of same dimension than `block_sizes` with the min
correlation for each block}

\item{epsilon}{maximum entry wise random noise}

\item{eidim}{dimension of the noise space, the smaller the value gets more
noise}
}
\value{
a simulated correlation matrix
}
\description{
Calculates a Toeplitz matrix with values descending from a user specified
range of correlations per block. The matrix has a block diagonal structure
and the size and base correlation for each block is user specified.
}
\note{
Based on Hardin et al, "A method for generating realistic
correlation matrices", Annals of Applied Statistics (2012)
}
\examples{
rho_hub = c(.9, .7)
tau_hub = c( (.9 - .7) / (10-2), (.7-.6)/(5-2))
eps_hub = min(1-(rho_hub) - 0.75*(tau_hub) ) - .01
sim_hub_cor(block_sizes = c(10, 5),
 max_corrs = rho_hub, min_corrs = c(.7,.6),
 power = 1, epsilon = eps_hub, eidim=2)
}
