% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_skew}
\alias{sim_skew}
\title{Simulated the allelic skew from a gaussian mixture with to components and
including a peak overlap event}
\usage{
sim_skew(n, p = 0.5, q = NULL, mean = 0, sd = 1)
}
\arguments{
\item{n}{number of SNPs}

\item{p}{probability of being in each component}

\item{q}{probability of don't overlapping a peak overlap}

\item{mean}{mean of the gaussian mixtures (one is N(mean, sd) and the other
is N(-mean, sd))}

\item{sd}{std. deviation of the allelic skew}
}
\value{
the simulated allelic skew for n SNPs
}
\description{
Simulated the allelic skew from a gaussian mixture with to components and
including a peak overlap event
}
\examples{
sim_skew(n = 20, p = .7, q = NULL, mean = 3, sd = 2)
sim_skew(n = 20, p = .7, mean = 3, sd = 2)
sim_skew(n = 20, p = .7, q = .8, mean = 3, sd = 2)
}
