% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{sim_toeplitz_cor}
\alias{sim_toeplitz_cor}
\title{Calculates an AR(1) Toeplitz matrix with a block diagonal structure. The size
and base correlation for each block is user specified}
\usage{
sim_toeplitz_cor(
  block_sizes = c(10, 5, 8, 2, 15, 50),
  corrs = c(0.7, 0.7, 0.5, 0.9, 0.85, 0.4),
  epsilon = 0.01,
  eidim = 2
)
}
\arguments{
\item{block_sizes}{A vector with the dimensions of each block}

\item{corrs}{A vector of same dimension than `block_sizes` with the base
correlation for each block}

\item{epsilon}{maximum entry wise random noise}

\item{eidim}{dimension of the noise space, the smaller the value gets more
noise}
}
\value{
a simulated correlation matrix
}
\description{
Calculates an AR(1) Toeplitz matrix with a block diagonal structure. The size
and base correlation for each block is user specified
}
\note{
Based on Hardin et. al., "A method for generating realistic
correlation matrices", Annals of Applied Statistics (2012)
}
\examples{
rho_top <- c(.7, .9)
eps_top = (1-max(rho_top))/(1+max(rho_top)) - .01
sim_toeplitz_cor(block_sizes=c(10, 5), corrs = rho_top,
 epsilon = eps_top, eidim = 2)
}
