% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_FMHighLD_simple}
\alias{simulate_FMHighLD_simple}
\title{Simulates a simple dataset with one causal per LD block}
\usage{
simulate_FMHighLD_simple(
  nsnps,
  nldblocks,
  ld_delta,
  ld_epsilon,
  ld_eidim,
  skew_p,
  skew_q,
  skew_mean,
  skew_sigma2,
  z_beta,
  z_sigma2
)
}
\arguments{
\item{nsnps}{Number of SNPs to simulate}

\item{nldblocks}{Number of LD blocks to simulate}

\item{ld_delta}{An additional parameter for the off diagonal correlations}

\item{ld_epsilon}{maximum entry wise random noise}

\item{ld_eidim}{dimension of the noise space, the smaller the value gets more
noise}

\item{skew_p}{probability of being in each component}

\item{skew_q}{probability of don't overlapping a peak overlap}

\item{skew_mean}{mean of the gaussian mixtures (one is N(mean, sd) and the
other is N(-mean, sd))}

\item{skew_sigma2}{variance of the allelic skew}

\item{z_beta}{regression coefficients of the z-statistic as function of the
annotation data}

\item{z_sigma2}{variance of the z-statistic as function of the annotation
data}
}
\description{
Simulates a simple dataset with one causal per LD block
}
\examples{
simulate_FMHighLD_simple(2000, 200, .2, .1, 2, .5, .9, 4, 1, c(0, 3), .5^2)
}
