% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_FMIter.R
\name{substitute_weights}
\alias{substitute_weights}
\title{Substitutes a column from the weight matrix}
\usage{
substitute_weights(i, weight, causal_candidate, gamma)
}
\arguments{
\item{i}{column to substitute into the weight matrix}

\item{weight}{the weight vector for every snp}

\item{causal_candidate}{the snp entries to substitute}

\item{gamma}{the posterior probability matrix}
}
\value{
a new vector of weights
}
\description{
Substitutes a column from the weight matrix
}
