% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{underlying_linear_model}
\alias{underlying_linear_model}
\title{performs the underlying linear model in FMHighLD}
\usage{
underlying_linear_model(i, i_data, formula, gamma, error_bound)
}
\arguments{
\item{i}{index of the model}

\item{i_data}{`fmld_data` for the causal candidates selected by the
underlying i-th model}

\item{formula}{a formula object with the underlying linear model}

\item{gamma}{estep probabilities}

\item{error_bound}{numerical constant to avoid zero eigenvalues}
}
\value{
an `lm` model
}
\description{
performs the underlying linear model in FMHighLD
}
