% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{underlying_lme_model}
\alias{underlying_lme_model}
\title{performs the underlying linear model in FMHighLD}
\usage{
underlying_lme_model(i, X, gamma, error_bound)
}
\arguments{
\item{i}{index of the model}

\item{X}{covariate matrix}

\item{gamma}{estep probabilities}

\item{error_bound}{numerical constant to avoid zero eigenvalues}
}
\value{
an `lmer` model
}
\description{
performs the underlying linear model in FMHighLD
}
