% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_causals.R
\name{causal_rule}
\alias{causal_rule}
\title{Applies the causal selection rule to the residual vector}
\usage{
causal_rule(data, residuals, fm_param, group, annot_names)
}
\arguments{
\item{data}{a `data.frame` used to fit the FM-HighLD model}

\item{residuals}{a vector of residuals for all the SNPs}

\item{fm_param}{configuration parameter used to pick the causal candidate
per group}

\item{group}{names of the grouping variable used to pick the causal
candidates}

\item{annot_names}{names of the annotation variables to ignore the variants
with `norm2(annot)` close to zero}
}
\description{
Applies the causal selection rule to the residual vector
}
