% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_iteration_single}
\alias{em_iteration_single}
\title{Performs an Expectation-Maximization iterations for the single
trait model}
\usage{
em_iteration_single(formula, data, pi, model_list, sigma0, verbose)
}
\arguments{
\item{formula}{a formula object with the underlying linear model}

\item{data}{a `data.frame` object with the annotation matrix and z-values}

\item{pi}{a vector of probabilities in the FMHighLD mixture part}

\item{model_list}{a list with the underlying linear models of FMHighLD}

\item{sigma0}{the error variance estimate of the background model}

\item{verbose}{a logical indicator of adding process messages}
}
\value{
a list with the updated parameters in one EM iteration
}
\description{
Performs an Expectation-Maximization iterations for the single
trait model
}
