% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_FMIter.R
\docType{methods}
\name{models}
\alias{models}
\alias{probmatrix}
\alias{models,FMIter-method}
\alias{probmatrix,FMIter-method}
\title{models method}
\usage{
models(object)

probmatrix(object)

\S4method{models}{FMIter}(object)

\S4method{probmatrix}{FMIter}(object)
}
\arguments{
\item{object}{a `FMIter` object}
}
\value{
a list of linear models for single trait or linear mixed models for
  the multitrait case

a `Matrix::Matrix` object with the posterior probabilities computed
in the latest EM iteration
}
\description{
`models` return the list of models estimated in the `FMHighLD` iteration

`probmatrix` returns the posterior probability matrix computed in the EM
algorithm latest iteration
}
