% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods_FMIter.R
\docType{methods}
\name{models}
\alias{models}
\alias{probmatrix}
\alias{sigma0}
\alias{causal_candidates}
\alias{prob_metric}
\alias{mccl}
\alias{coef_diff}
\alias{coef_diff_em}
\alias{loglikelihood}
\alias{get_causal_data}
\alias{models,FMIter-method}
\alias{probmatrix,FMIter-method}
\alias{sigma0,FMIter-method}
\alias{causal_candidates,FMIter-method}
\alias{prob_metric,FMIter,FMIter-method}
\alias{loglikelihood,FMIter,data.frame-method}
\alias{mccl,FMIter,FMIter-method}
\alias{coef_diff,FMIter,FMIter,logical,logical-method}
\alias{coef_diff_em,FMIter,FMIter,logical,logical-method}
\alias{get_causal_data,FMIter,data.frame-method}
\title{models method}
\usage{
models(object)

probmatrix(object)

sigma0(object, response)

causal_candidates(object)

prob_metric(object, prev)

mccl(object, prev)

coef_diff(object, prev, model_error, background_error)

coef_diff_em(object, prev, model_error, background_error)

loglikelihood(object, fmld_data)

get_causal_data(object, fmld_data)

\S4method{models}{FMIter}(object)

\S4method{probmatrix}{FMIter}(object)

\S4method{sigma0}{FMIter}(object, response)

\S4method{causal_candidates}{FMIter}(object)

\S4method{prob_metric}{FMIter,FMIter}(object, prev)

\S4method{loglikelihood}{FMIter,data.frame}(object, fmld_data)

\S4method{mccl}{FMIter,FMIter}(object, prev)

\S4method{coef_diff}{FMIter,FMIter,logical,logical}(object, prev, model_error, background_error)

\S4method{coef_diff_em}{FMIter,FMIter,logical,logical}(object, prev, model_error, background_error)

\S4method{get_causal_data}{FMIter,data.frame}(object, fmld_data)
}
\arguments{
\item{object}{The `FMIter` object}

\item{response}{the vector of all the z-values in the FMHighLD model}

\item{prev}{The previous `FMIter` iteration of the object}

\item{model_error}{a logical value indicating whether including the model
errors' difference to the output vector}

\item{background_error}{a logical value indicating whether including the
background errors' difference to the output vector}

\item{fmld_data}{A `data.frame` with the response and annotation matrix}
}
\value{
a list of linear models for single trait or linear mixed models for
  the multitrait case

a `Matrix::Matrix` object with the posterior probabilities computed
in the latest EM iteration

the std. deviation of the non-causal mixture

a list of causal candidates for every mixture component

a vector with the jsd metric for every mixture components

the mean one-zero loss of causal candidates variants between the
 current and previous iteration

the distance between the fixed effect coefficients

the distance between the fixed effect coefficients

The log-likelihood of the FMHighLD model

`fmld_data` for the causal candidates
}
\description{
`models` return the list of models estimated in the `FMHighLD` iteration

`probmatrix` returns the posterior probability matrix computed in the EM
algorithm latest iteration

`sigma0` obtains the std. deviations of the non-causal components

Gets lists of causal candidates for every mixture
}
